Update patch set 1
Patch Set 1: Code-Review+1

No objection to this change, but I think Tobias is right. A revert + reapply is safer. In this case, even more so, because the test line doesn't indicate a user build was tested.

I'm going to +1 this change, cuz I like it, but I +2d the revert in the meantime.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1