Update patch set 4
Patch Set 4:

(1 comment)

I agree that SEPOLICY_SPLIT is the smaller component.  My concern is that once we define it as its own flag, it disconnects from its original meaning.  Another way of viewing this question is, where is it set and how?  The sepolicy is meant only to reflect what's happening on devices, so any conditional inclusion/exclusion of policy should be grounded in architectural decisions and I'm not sure what decision PRODUCT_SEPOLICY_SPLIT will reflect.

Patch-set: 4