Update patch set 1
Patch Set 1:

Ah ha, I see what's going on here.

This comes down to the difference between setgit bit on dirs and files. Because you have the setgid bit on the directory, the new file is correctly inheriting the GID. However, CAP_FSETID is checking to see if the new file should also have the setgid bit set, which it should not since that's used for putting processes into the correct group upon exec which is not the intent here. Similar to aosp/145219, this seems like a good candidate for dontaudit.

Patch-set: 1
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>