WIP: Allow SysApps to Receive UDP Sockets from SystemServer
This gives the privilege to system apps to receive a
UDP socket from the system server. This is being added
for supporting UDP Encapsulation sockets for IPsec, which
must be provided by the system.

This is an analogous change to a previous change that
permitted these sockets for untrusted_apps:
0f75a62e2c4fb1b6ef8db6f2e5c10ff29f95322d

Bug: 70389346
Test: WIP
Change-Id: Iec07e97012e0eab92a95fae9818f80f183325c31