SELinux policies for Perfetto cmdline client (/system/bin/perfetto)
Instead of having statsd linking the perfetto client library
and talk directly to its socket, we let just statsd exec()
the /system/bin/perfetto cmdline client.

There are two reasons for this:
1) Simplify the interaction between statsd and perfetto, reduce
  dependencies, binary size bloat and isolate faults.
2) The cmdline client also takes care of handing the trace to
  Dropbox. This allows to expose the binder interaction surface
  to the short-lived cmdline client and avoid to grant binder
  access to the perfetto traced daemon.

The trace file is created unlinked (using O_TMPFILE) and is
linked on the /data/misc/perfetto-traces filesystem only to
handle it to Dropbox, and immediately re-unlinked. The relevant
code lives here:
https://cs.corp.google.com/android/external/perfetto/src/traced/perfetto_cmd/perfetto_cmd.cc?l=285

This cmdline client will be used by:
 - statsd
 - the shell user (for our UI and Studio)

Bug: 70942310
Change-Id: I8cdde181481ad0a1a5cae5937ac446cedac54a1f