Update patch set 1
Patch Set 1:

> I suspect that this is happening because installd is opening /proc/mounts without O_CLOEXEC, so the FD gets passed to idmap, even though it doesn't need it.

That was my thought, except that the fd is only used locally within that function (and only within calls to getline).  So I was thinking perhaps it's a problem with multiple threads and the fork?

But yes, I'll look into this tomorrow.

Patch-set: 1
Reviewer: Gerrit User 1208973 <1208973@85c56323-6fa9-3386-8a01-6480fb634889>