Allow perfetto traced_probes to access tracefs on user
Allows the traced_probes daemon to access the core ftrace
functionalities on user builds. Specifically this involves:
- Whitelisting the trace_pipe_raw file to allow perfetto to
  read the raw ftrace buffer (rather than the text-based
  /trace endpoint)
- Whitelisting the /stats and /buffer_size_kb that allow to
  tune the buffer size per-cpu pipe and to get basic statistics
  about the ftrace buffer (#events captured, overrun)
- Whitelistiing the /format files for the events enabled on user
  builds. "format" files simply describe the memory layout
  of the binary logs. Example: https://ghostbin.com/paste/f8m4k

This still does NOT allow enabling the events labeled as
"_debug" (mostly events that return activity on inodes).
We'll deal with that separately as soon as we get a POC
of inode resolution and a sensible blacklist/whitelist model.

Bug: 70942310
Change-Id: Ic15cca0a9d7bc0e45aa48097a94eadef44c333f8