Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> How long must we support both methods of launching webview_zygote?

We don't intend to keep both. As soon as the topic at ag/3535340 lands we can remove the old launching path. However since that's not landing in AOSP I was wondering how to proceed with the SELinux cleanup: doing it in AOSP would mean breaking public master, so I was thinking of doing that in ag/. Does that sound okay?

Patch-set: 1
Reviewer: Gerrit User 1042203 <1042203@85c56323-6fa9-3386-8a01-6480fb634889>