Update patch set 1
Patch Set 1:

I understand that the backing file looks after it. I also understand that this is a _minor_ issue that the layers of security look after anyways.

The subtle difference is that when I tried to use shell-completion it _did_ complete mkfs.ext[234], but did not for mke2fs and that was when I decided pedantically we need to "protect" the symbolic links too. This makes the protection consistent.

If we decided to save space on our filesystems, we may introduce hard links rather than symbolic links, and if we do so we _must_ have the labels for each presented directory node.
 This is true for _all_ toybox symbolic links too. A move to hard links will also improve security. For example we are giving way too much to toybox, and not refined enough for the individual executables. Complexity could be costly though ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>