Useful neverallow errors
Neverallow rules are include the file and line number of the
violated rule. However, if a number of neverallow rules are
contained within a macro, then the line number is for the macro,
not the individual neverallow rule that is violated. To fix this,
only include one neverallow rule per macro.

This changes nothing on device, nor does it change the results of
compilation, it only makes the errors printed more useful.

Bug: 69139821
Test: build aosp_taimen-userdebug (neverallow rules are build time
    tests)

Change-Id: Id0fc5906431db20e71265c7e9d55fbee4bdf53db