Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Yes, isn't there an interface in liblogcat for this purpose?  create_android_logcat() / android_logcat_run_command() / etc?

Agreed, but with some reality checks or provisos:

We stopped short of actually _using_ it on dumpstate.cpp for the upcoming release (which would have served as a great example) because there is no timeout facility (yet) should the in-process liblogcat stall for too long. It should be trivial to add a timeout facility for the liblogcat API. Of course, timeouts on logcat executable is extremely rare, and only really happen with the target is in a very sick state already; so an initial spin by the partner to use it should be fine.

There has been no discussion to export liblogcat, but it does use a "C" interface, so it should be sustainable/maintainable for use by the partners. One of the problems not addressed on 64-bit devices is we only land the 64-bit version of liblogcat.so, nothing on /system requests the 32-bit version of liblogcat.so to make it truly usable under all conditions unless we make sure "both" get landed on /system.

Faster/efficient to use liblogcat.so instead of pipe/fork/exec("logcat"...)

Please file a bug to add a "C" method to liblogcat.so that references the context to land a timeout to address this shortcoming so we can finally deploy it internally everywhere. We (Android) also need to finish b/28035297 (system_server to utilize liblogcat.so via JNI) and https://android-review.googlesource.com/c/platform/system/core/+/342474/.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>