Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > There has been no discussion to export liblogcat, but it does use a "C" interface, so it should be sustainable/maintainable for use by the partners
> 
> Do you mean that we don't have it as part of the VNDK yet?

The library is not turned on for vndk. I believe since it is an integral part of liblog+logd support, that it belongs in a special place, the same place bionic libc and liblog.so currently fill. It has dependencies on other libraries like libbase ...

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>