Update patch set 2
Patch Set 2:

this is used by the legacy HAL implementation in libhardware which is used by the legacy audio HAL wrapper.
I don't recall the full history but it was probably for OEMs to override the name of the HAL .so library from the default naming convention.

Patch-set: 2
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>