Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> this is used by the legacy HAL implementation in libhardware which is used by the legacy audio HAL wrapper.
> I don't recall the full history but it was probably for OEMs to override the name of the HAL .so library from the default naming convention.

Then we should make it a vendor property (ro.vendor) and not add it to the white list, right? If the platform doesn't use it, we should constrain it to the vendor partition only. I understand this is probably the case for other properties on this list, and their presence here might make this distinguishing feature less than clear, but doing the right thing moving forward is what I'm after. Jaekyun can comment on this (added) or someone more familiar with the (actual implementation of the) system/vendor property split might also comment.

Patch-set: 2
Reviewer: Gerrit User 1019198 <1019198@85c56323-6fa9-3386-8a01-6480fb634889>