Update patch set 2
Patch Set 2:

There is no use of this property directly in the code. I'm told it's empty. I'm assured it's only used for a system<->system binary discovery. Those three things tell me there's no reason for it to be added to this list.

Why not just make it a system-only property?

Patch-set: 2