Update patch set 2
Patch Set 2:

> Can you explain what you mean by "system<->system binary discovery" in this context?

I think he means that the system should not have to discover the system image at run time but rather at build time. Thus this property should not be used.

> Why not just make it a system-only property?

It is only read by the system right now, and will only be read by the vendor when A2DP will finally be binderized.
This property is an old leftover of the libhardware way of loading libraries still in use today because the audio HALs are still packaged in libhardware modules.

I think it is better to stay consistent with the other properties or do the correct thing and switch all properties to their respective partition.
Nevertheless, I stand for doing that in a separate patch.

Patch-set: 2
Reviewer: Gerrit User 1136369 <1136369@85c56323-6fa9-3386-8a01-6480fb634889>