Update patch set 1
Patch Set 1:

(1 comment)

FWIW, if there are code changes in netd that we can make to avoid these denials, I'd be fine with them. That may be a better option here.

Also, is this due to a kernel bug where the permissions checks are done in the wrong order? If so, can we fix that bug, or are we tracking it somewhere?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>