Update patch set 9
Patch Set 9: Presubmit-Verified-1

=== Finished presubmit run: L72700000155015757 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/635599?revision=9&workplanId=L72700000155015757




TreeHugger finished with: 16 passed, 1 failed, 1 skipped.





WARNING: The tip-of-tree(4669915) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4669418), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







WARNING: The tip-of-tree(4669915) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4669525), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4669915) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4669616), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




FAIL: Failed to build git_master/taimen-userdebug/4669915 (ToT). See https://android-build.googleplex.com/builds/pending/P5586264/taimen-userdebug/latest



WARNING: The tip-of-tree(4669915) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4669501), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 9
Tag: autogenerated:TreeHugger:Verify-Fail
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=-1