added attribute "persist_dir_file_write_violators"
moving the persist mount point to /vendor will force it to
make /vendor/persist  as vendor_file_type which is read only.

But as persist will have initial setupfile, config files
that change during the runtime like sensors/display driver
and regionalization data is been update to this location .
Which involves write/create/unlink sort of file operation .

So using of this violator is going to enable this.
Which should be deprecated  once we have cleaner way .
b/74925484

Change-Id: I4591cbf243fa307109e0563c2a124b558a830225