Update patch set 1
Patch Set 1:

A number of capabilities checks occur just because the process is running as root, but which we do not want to grant, and can safely deny. For cases like that where the denial is WAI, it's better to not spam the logs with selinux denials which cause red herrings for other teams.

Patch-set: 1
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>