Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can we not use links?

If we don't use links, we'd have to create temporary files and rename them in (and have to deal with subsequent problems like stale files remaining on a reboot). If the objection is around having hard links on the filesystem in general, the link count on the file being linked is 0.

Patch-set: 1
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>