Update patch set 1
Patch Set 1: Presubmit-Verified-1

=== Finished presubmit run: L27800000167948807 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/674570?revision=1&workplanId=L27800000167948807




TreeHugger finished with: 17 passed, 1 failed, 2 skipped.





WARNING: The tip-of-tree(4750044) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4749905), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.









WARNING: The tip-of-tree(4750044) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4749905), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





WARNING: The tip-of-tree(4750044) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4749928), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




FAIL: Failed to build git_master/taimen-userdebug/4750044 (ToT). See https://android-build.googleplex.com/builds/pending/P5863466/taimen-userdebug/latest



WARNING: The tip-of-tree(4750044) for git_master/walleye-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4749875), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 1
Tag: autogenerated:TreeHugger:Verify-Fail
Reviewer: Gerrit User 1062513 <1062513@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=-1