Update patch set 3
Patch Set 3: Code-Review+1

I like this change, doesn't do what we eventually want, but its the best we can go with right now to partly close the same_process_hal_file loophole.

If/when, we have a way to ensure the coredomain apps are only coming from platform (i.e. /system), I think I can make an argument for extending this rule to coredomain apps too.

If there is a coredomain app that relies on a vendor_public_library, there is an argument to be made to either - 

1. Add that library in /system so now the app can use it but don't remove the vendor public library so untrusted apps can continue to rely on it.

OR

2. (better alternative) move the app to /vendor. Why should systemui depend on a vendor library? (from the example cited above).

Patch-set: 3
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1