Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I confirmed that the tests that work on v4.9 don't work with v4.14 kernels. Since, we do want to make sure they continue to work for 4.14 as well, This seems like the quickest way to work around the problem until we reach a resolution on what changed in the kernel and see if that has to be amended or not.
> 
> I must say, that needing a write permission for 'kernel' domain for loopback mounts doesn't sound right to me, since those write / reads are only on behalf of userspace domain (in this case 'su'). We know the kernel needed 'read' before for losetup and attaching /dev/loopX to a filesystem image and that sounds reasonable. I still don't fully understand why 4.14+ kernels started requiring 'write' access as well.

As mentioned in the seandroid-list thread cited above, this appears to be due to a vfs refactoring of vfs_iter_write() that causes it to reuse a code path that calls rw_verify_area(), and this calls security_file_permission().  Either it needs to not do that, or security_file_permission() needs to be moved out of rw_verify_area() and taken to the callers where it is appropriate.  Other model would be to save the cred somewhere and override the current cred with that cred during these calls so that access is checked based on the userspace entity that performed the mount.

Patch-set: 1
Reviewer: Gerrit User 1010111 <1010111@85c56323-6fa9-3386-8a01-6480fb634889>