Update patch set 6
Patch Set 6:

I don't think we'd want to use a script to generate it for auditing reasons.  

We could have an attribute that we apply to all of the platform properties, but that seems to be a direction that we haven't really been going.

I'd defer to jeffv@ on how he thinks this would be best done.

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>