Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> > Patch Set 6:
> > 
> > I don't think we'd want to use a script to generate it for auditing reasons.  
> > 
> > We could have an attribute that we apply to all of the platform properties, but that seems to be a direction that we haven't really been going.
> > 
> > I'd defer to jeffv@ on how he thinks this would be best done.
> 
> Me neither. Given that we would like to (gradually) kill sysprops, I don't expect we will add much to this list and any addition should be audited.

The way we've handled this for the other Treble-split types is to introduce a new attribute for marking core types which can be added to the type declaration. We already have that for properties: core_property_type. Instead of adding to this list, just require that properties owned by coredomains have the core_property_type attribute.

Patch-set: 6
Reviewer: Gerrit User 1054468 <1054468@85c56323-6fa9-3386-8a01-6480fb634889>