Update patch set 2
Patch Set 2:

That would require a restart of the device which happens possibly once a month. The update of the apns is done when the device is already running, the functionality of writing the file is already placed in system server/ConfigUpdateInstallReceiver. It broke with latest sepolicy changes

Patch-set: 2
Reviewer: Gerrit User 1221112 <1221112@85c56323-6fa9-3386-8a01-6480fb634889>