Update patch set 2
Patch Set 2:

> Why not leave files in the initramfs version of /system as rootfs?

If we don't relabel these and thus they are labeled as rootfs then we need to add many recovery-only rules. For example, without this change, `adb shell` is not working because adbd is not allowed to execute a file of type rootfs.

> Requiring that "system_file" be a build-time-only label has security benefits

In general I agree, but the extent of this change is very limited. Only the init process running in the recovery mode is allowed to do the relabel from 'rootfs' to 'system_file'. This does not affect other processes in the recovery mode not to mention the normal mode. Even the recovery init is not allowed to use this to relabel things in the persistent storages (e.g. the real system.img) because they are mounted read-only and we even have a plan to not mount the image at all in recovery mode. So, I don't understand what bad result could happen with this change.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>