Update patch set 3
Patch Set 3:

(1 comment)

> So why are you adding:
>     allow init adbd_exec:file relabelto;
> ?

That's not strictly needed (at this moment). I've removed it. 

But with that, we can instead remove the recovery-specific rules that I've shown above. One of the goals of this topic is to make the recovery partition as close as the system partition (e.g. directory layout, etc.). So why not for sepolicy? It's unfortunate that rootfs does not support xattr (I see there is on-going effort for that but it seems that it's not officially supported yet). If it had supported xattr, we wouldn't need these rules for relabeling (which themselves are recovery-specific, ironically) at all.

> CDD says that anything with the Linux kernel must be in selinux enforcing mode. That certainly applies to recovery (and charging mode, etc.), and if vendors want to use a separate kernel, they must create an equivalent mandatory access control system.

Well, that's right. But I think using this to relabel persistent partitions will eventually blocked by the verified boot chain, because the partition should have been invalidated by the relabeling.

Anyway, do you have better idea in your mind?

Patch-set: 3
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>