Update patch set 1
Patch Set 1: Code-Review-1

API council, not yet (had to skip the last three active one b/c of appointments and interviews) and have lacked the bravery to bring it up otherwise, so my bad.

This is the "Canonical Boot Reason", sys.boot.reason was meant to be the API and has a strict format described in b/65686279 for the CDD for P. For the 'system' really only init (rc files), dumpstate and Power Manager (which forms an API for reporting a thermal shutdown). For 'vendor' on the other hand, we have received a couple of emails so far requesting access so that HALs can make decisions based on why the system rebooted, they dutifully add it to their sepolicy and access it. I expect more.

Let this CL rest until I get approval ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1