Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> I really don't think that's how it works. The linker should be called after the exec which is also after the domain transition has occurred. I'll look into this.

I've re-read the kernel code and what I said before is not fully correct, but the result is the same.

load_elf_binary() @ fs/binfmt_elf.c

The function first opens the program interpreter (which is the dynamic linker) for execution. open_exec(elf_interpreter);

Then it commits to the new domain via install_exec_creds(bprm). So, at the moment when the linker is opened for execution, the process is still in the previous domain. (But yes it's well before the loading and linking are performed. I was wrong for this.)

Patch-set: 6
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>