Update patch set 1
Patch Set 1: Code-Review-1

This CL is really about sepolicy_test.py changes needing a security opinion as discussed in the hallway.

Incomplete as it does not provide any MAC for adb (userdebug systems, verity off/on or remount, all after adb root) in order to create the directory tree in /cache/overlay/* and/or /data/overlay/*. The rules for adb may reside in a separate dependent CL, or may merge into this one. 

Basically, if the /data is mounted _when_ system is mounted, in mount_all --early, then /data can be used for the upperdir for overlayfs, if not, then /cache (if available, non-a/b systems) may be used. nothing encrypted just as it is unencrypted on /system or /vendor etc. For later mount of /data and no /cache (a/b systems), then other arrangements need to be taken of course ...

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1