Update patch set 1
Patch Set 1:

> No, but the remount is done from the security context of su, which always runs in selinux permissive mode. Now that I have a better idea of this change, perhaps we should meet to discuss.

The normal readonly mount of overlayfs during reboot after adb disable-verity, however, is done in the context of init during mount_all. The _emergency_ _backup_ mount of overlayfs, if it is not currently mounted, I propose in adb remount would be in the su context.
 
> > I got 'confused' by how asan builds were handled, I will re-try figuring it out as maybe I was making this more complex than it was. From what you say it sounds as simple as placing the new changes into private/file_contexts_overlayfs.

For some reason, even though I can see the contents of file_contexts_overlayfs in /plat_file_contexts on the device, restorecon -R /data/overlay/ is always reporting system_data_file default and the rules do not appear to be applied. Is there an ordering issue I need to be aware of? (I have tried to concatenate file_contexts_overlayfs before and after with no joy). But in place as in this CL in file_contexts, it works?

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>