Allow apps with the DUMP permission to use battery service dumpsys
"dumpsys battery" contains the following output:
  AC powered: false
  USB powered: true
  Wireless powered: false
  Max charging current: 900000
  Max charging voltage: 5000000
  Charge counter: 3518290
  status: 2
  health: 2
  present: true
  level: 100
  scale: 100
  voltage: 4358
  temperature: 254
  technology: Li-ion

All the information is available from elsewhere, but in a "cooked" form.

Sometimes it's very useful to be able to see the "raw" numbers, without
connecting a device to a PC.

This CL allows apps with the DUMP permission to use "dumpsys battery".
(which is how we protect dumpsys of many other services.)

The battery service has only one other binder method, onShellCommand(),
which already has a proper caller check.

Bug: 110088132
Test: Manual test with go/dumpviewer
Change-Id: Ic170a4495e11dd88eeae91cb888dd466caab6ff5