Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Here is one question. Denying for every secondary dex file would not be what you want. In fact, I don't know the exact meaning "don't allow to open, but allow to read". Can I know what is the content of the specified "Bug: 67111829"?

This was done to strengthen the file access model at the request of the security team. The intention is to only pass (opened) file descriptors to dexoptanalyzer. It's a similar to all the other installd/dexopt operations: installd opens the file and then passes the descriptor along; as such the tool need only read permissions without the ability to open arbitrary file names.

Patch-set: 1
Reviewer: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>