Update patch set 1
Patch Set 1:

(1 comment)

> This change would mean we support kernel module loading from two different places, ueventd and modprobe. See, for example https://android.googlesource.com/platform/system/sepolicy/+/master/public/modprobe.te

> I'm nervous having module loading which occurs in two different places in the code tree. Kernel module loading is arbitrary code execution in the kernel, probably about the most dangerous thing imaginable.  Having this occurring in two different places without a good explanation feels ... scary, both from a security perspective as well as a reliability perspective. Can we delete one of these?

Probably not.  They serve two different use cases.  Modprobe is good for devices that already know which specific modules they're going to load whereas this mechanism is made for modular boards that know they will load some modules from a given set of modules, but not necessarily all.  We had a conversation on the change that added this functionality to ueventd and decided that modprobe is not sufficient/appropriate.

I do plan to make this opt into, which answers your second question too.

> Second question: Do we expect module loading to be common on all devices (so we should check it into common policy), or do we expect it to be device specific (which would imply we should check it into device-specific policy)?

No it won't be common across all devices.  We can abandon this and have vendors put it in their own policy.  That seems a bit clunky to me that we'd have the mechanism in the platform ueventd but require vendors to have the permissions in their device specific sepolicy.  Though I guess it's not much different than init, where insmod is supported but vendors need to put the permissions in their device specific sepolicy.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0